chore: create routes package with centralized destinations
1. Remove routes package and Destinations.kt file
2. Restore navigation destinations to their original screen files
3. Remove imports for routes package
