+++ app/src/main/java/com/example/mycontacts/presentation/authentication/LoginPage.kt
@@ -23,12 +23,6 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.mycontacts.navigation.NavigationDestination
-
-
-object LoginPageDestination : NavigationDestination {
-    override val route = "login"
-}
 
 @Composable
 fun LoginPage(
