+++ app/src/main/java/com/example/mycontacts/presentation/contact/detailScreen/DetailsScreen.kt
@@ -17,14 +17,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.mycontacts.ContactsTopAppBar
 import com.example.mycontacts.R
-import com.example.mycontacts.navigation.NavigationDestination
+import com.example.mycontacts.navigation.routes.DetailsScreenDestination
 import kotlinx.coroutines.launch
-
-object DetailsScreenDestination : NavigationDestination {
-    override val route = "contact_details"
-    const val contactIdArg = "contactId"
-    val routeWithArgs = "$route/{$contactIdArg}"
-}
 
 @Composable
 fun DetailsScreen(
