+++ app/src/main/java/com/example/mycontacts/presentation/contact/detailScreen/DetailsScreenViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.mycontacts.presentation.contact.ContactUiState
 import com.example.mycontacts.domain.repository.ContactsRepository
+import com.example.mycontacts.navigation.routes.DetailsScreenDestination
 import com.example.mycontacts.presentation.contact.toContact
 import com.example.mycontacts.presentation.contact.toContactUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
