+++ app/src/main/java/com/example/mycontacts/presentation/contact/editScreen/EditScreenViewModel.kt
@@ -11,6 +11,7 @@ import com.example.mycontacts.presentation.contact.isValid
 import com.example.mycontacts.presentation.contact.toContact
 import com.example.mycontacts.presentation.contact.toContactUiState
 import com.example.mycontacts.domain.repository.ContactsRepository
+import com.example.mycontacts.navigation.routes.EditScreenDestination
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
